+++ src/api/api.ts
@@ -9,6 +9,7 @@ import {
 } from '@/src/types/api.types';
 import { encodeIngredients } from '@/src/helpers/query.helper';
 import { UserData } from '@/src/types/register.types';
+import { ApiError } from 'next/dist/server/api-utils';
 
 export async function getMeal(id: string): Promise<Meal> {
     return apiGet<Meal>(`meals/${id}/details`);
@@ -80,6 +81,12 @@ export async function confirmUserActivation(login: string): Promise<void> {
     await apiPost(`users/${login}/activate`);
 }
 
-export async function createUserAccount(data: UserData) {
-    await apiPost<UserData>('/users/create', data);
+export async function createUserAccount(data: UserData): Promise<void> {
+    const res = await apiPost<UserData>('users/create', data);
+
+    if (res.status > 299) {
+        const json = await res.json();
+
+        throw new ApiError(res.status, json.message);
+    }
 }